improve input selection
input selection now works the same way as in the python impl - if we
encounter a filter, we choose our values from the lhs of the filter.
this improves rankings for some benches and fixes a bug in RE.

additionally, if no input values are available for a type, just
return None. This fixes another bug in RE.